+++ routes/api/contactSchema.js
@@ -1,0 +1,11 @@
+const Joi = require('joi');
+
+const contactSchema = Joi.object({
+  name: Joi.string().min(3).max(30).required(),
+  email: Joi.string().email().required(),
+  phone: Joi.string()
+    .pattern(/^\d{10}$/)
+    .required(),
+});
+
+module.exports = contactSchema;
